third_party/blink/web_tests/external/wpt/css/css-grid/masonry/tentative/track-sizing/auto-repeat/row-auto-repeat-023-expected.txt

This is a testharness.js-based test.
[FAIL] Auto-fit repeat tracks with fixed size and auto item placement
  assert_equals: expected "100px 100px 100px 100px 100px 100px 100px 100px 100px 100px" but got "repeat(auto-fit, 100px) repeat(5, 100px)"
Harness: the test ran to completion.
